+++ crates/rpc/src/schema_registry/types.rs
@@ -22,7 +22,7 @@ rpc_enum! {
         Bool,
         String,
         Integer,
-        Decimal,
+        Float,
         Any
     ]
 }
